refactor formatter
